refactor: modalDrawer 패키지 옮기고, UI 수정
