+++ presentation/src/main/kotlin/land/sungbin/androidprojecttemplate/home/HomeScreen.kt
@@ -5,9 +5,7 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -15,15 +13,12 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.ModalBottomSheetValue
 import androidx.compose.material.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.NonRestartableComposable
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.DpSize
@@ -40,6 +35,7 @@ import land.sungbin.androidprojecttemplate.common.UiStatus
 import land.sungbin.androidprojecttemplate.common.component.DuckieLoadingIndicator
 import land.sungbin.androidprojecttemplate.domain.model.Feed
 import land.sungbin.androidprojecttemplate.domain.model.constraint.FeedType
+import land.sungbin.androidprojecttemplate.home.component.DrawerContent
 import land.sungbin.androidprojecttemplate.home.component.DuckDealHolder
 import land.sungbin.androidprojecttemplate.home.component.FeedHeader
 import land.sungbin.androidprojecttemplate.home.component.FeedHolder
@@ -48,19 +44,12 @@ import land.sungbin.androidprojecttemplate.home.component.HomeNormalFeed
 import land.sungbin.androidprojecttemplate.home.component.dummyTags
 import land.sungbin.androidprojecttemplate.home.component.getTradingMethodResourceId
 import land.sungbin.androidprojecttemplate.home.component.priceToString
-import land.sungbin.androidprojecttemplate.shared.compose.extension.noRippleClickable
-import team.duckie.quackquack.ui.component.QuackBody2
 import team.duckie.quackquack.ui.component.QuackBottomSheetItem
-import team.duckie.quackquack.ui.component.QuackDivider
-import team.duckie.quackquack.ui.component.QuackHeadLine2
 import team.duckie.quackquack.ui.component.QuackHeadlineBottomSheet
 import team.duckie.quackquack.ui.component.QuackImage
 import team.duckie.quackquack.ui.component.QuackMenuFabItem
 import team.duckie.quackquack.ui.component.QuackModalDrawer
-import team.duckie.quackquack.ui.component.QuackRoundImage
 import team.duckie.quackquack.ui.component.QuackSimpleBottomSheet
-import team.duckie.quackquack.ui.component.QuackTitle1
-import team.duckie.quackquack.ui.component.QuackTitle2
 import team.duckie.quackquack.ui.component.QuackTopAppBar
 import team.duckie.quackquack.ui.component.rememberQuackDrawerState
 import team.duckie.quackquack.ui.icon.QuackIcon
@@ -267,7 +256,7 @@ fun HomeComponent(
                     FeedType.Normal -> {
                         HomeNormalFeed(
                             FeedHolder(
-                                profile = feed.writerId,
+                                profile = R.drawable.duckie_profile,
                                 nickname = feed.writerId,
                                 time = "3일 전", //Date 로직 작성 필요
                                 content = feed.content.text,
@@ -293,7 +282,7 @@ fun HomeComponent(
                     FeedType.DuckDeal -> {
                         HomeDuckDealFeed(
                             FeedHolder(
-                                profile = feed.writerId,
+                                profile = R.drawable.duckie_profile,
                                 nickname = feed.writerId,
                                 time = "3일 전", //Date 로직 작성 필요
                                 content = feed.content.text,
@@ -341,192 +330,4 @@ fun HomeComponent(
         },
         paddingValues = homeFabPadding,
     )
-}
-
-@Composable
-fun DrawerContent() {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(
-            space = 20.dp,
-        )
-    ) {
-        DrawerHeader()
-        QuackDivider()
-        Column(
-            modifier = Modifier.padding(
-                start = 16.dp,
-            ),
-            verticalArrangement = Arrangement.spacedBy(
-                16.dp,
-            )
-        ) {
-            QuackTitle2(
-                text = stringResource(id = R.string.trading_activity)
-            )
-            DrawerIconText(
-                icon = QuackIcon.Heart,
-                text = stringResource(id = R.string.list_of_interests),
-                onClick = {
-
-                }
-            )
-            DrawerIconText(
-                icon = QuackIcon.Sell,
-                text = stringResource(id = R.string.sales_history),
-                onClick = {
-
-                }
-            )
-            DrawerIconText(
-                icon = QuackIcon.Buy,
-                text = stringResource(id = R.string.purchase_history),
-                onClick = {
-
-                }
-            )
-        }
-        QuackDivider()
-        Column(
-            modifier = Modifier.padding(
-                start = 16.dp,
-            ),
-            verticalArrangement = Arrangement.spacedBy(
-                16.dp,
-            )
-        ) {
-            QuackTitle2(
-                text = stringResource(id = R.string.my_activity)
-            )
-            DrawerIconText(
-                icon = QuackIcon.Area,
-                text = stringResource(id = R.string.setting_up_areas_interest),
-                onClick = {
-
-                }
-            )
-            DrawerIconText(
-                icon = QuackIcon.Tag,
-                text = stringResource(id = R.string.edit_interest_tags),
-                onClick = {
-
-                }
-            )
-        }
-        QuackDivider()
-        DrawerIconText(
-            modifier = Modifier.padding(
-                start = 16.dp,
-            ),
-            icon = QuackIcon.Setting,
-            text = stringResource(id = R.string.app_settings),
-            onClick = {
-
-            }
-        )
-    }
-}
-
-@Composable
-private fun DrawerHeader() {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(
-                vertical = 20.dp,
-                horizontal = 16.dp,
-            ),
-        verticalArrangement = Arrangement.spacedBy(
-            space = 12.dp,
-        )
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .noRippleClickable {
-
-                }
-                .padding(
-                    vertical = 4.dp,
-                ),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(
-                    space = 16.dp,
-                )
-            ) {
-                QuackRoundImage(
-                    src = R.drawable.duckie_profile,
-                    size = DpSize(
-                        width = 48.dp,
-                        height = 48.dp,
-                    )
-                )
-                QuackHeadLine2(
-                    text = "닉네임"
-                )
-            }
-            QuackImage(
-                src = QuackIcon.ArrowRight,
-            )
-        }
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(
-                space = 42.dp,
-            )
-        ) {
-            DrawerNumberText(number = "2.6만", text = stringResource(id = R.string.follower))
-            DrawerNumberText(number = "167", text = stringResource(id = R.string.following))
-            DrawerNumberText(number = "88", text = stringResource(id = R.string.feed))
-        }
-    }
-
-}
-
-@Composable
-private fun DrawerNumberText(
-    number: String,
-    text: String,
-) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(
-            2.dp,
-        )
-    ) {
-        QuackTitle2(
-            text = number
-        )
-        QuackBody2(
-            text = text,
-        )
-    }
-}
-
-@NonRestartableComposable
-@Composable
-private fun DrawerIconText(
-    modifier: Modifier = Modifier,
-    icon: QuackIcon,
-    text: String,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier = modifier
-            .noRippleClickable(
-                onClick = onClick,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(
-            space = 8.dp,
-        )
-    ) {
-        QuackImage(
-            src = icon,
-        )
-        QuackTitle1(
-            text = text
-        )
-    }
 }
